refactor: place MSplash in you own folder
