+++ src/components/layouts/MSplash/MSplash.spec.ts
@@ -1,6 +1,6 @@
 import { testComponentFactory } from '@/utils/testComponentFactory';
 import { describe, expect, it } from 'vitest';
-import MSplashVue from './MSplash.vue';
+import MSplashVue from '@/components/layouts/MSplash/MSplash.vue';
 
 describe('Layouts: MSplash', () => {
 	it('should render the component correctly', () => {
