+++ src/components/layouts/MSplash/MSplash.vue
